refactor: trocear PreviewChannelModal en components/preview-channel (Fase 1.9)
847 → 349 líneas. Piezas decorativas del "Laboratorio Alpha" extraídas a
components/preview-channel/, cada una con sus estilos y constantes:
AnimatedGradients, FloatingParticle, ConfettiBurst, GiantLever, LabDecorations.

Cero cambios de comportamiento (mismo código, estilos repartidos sin
solapamiento); solo app/_layout.tsx consume el modal y su import no cambia.
Gigantes >800: 13 → 12. typecheck/typecheck:tests/lint(0 err)/test(183) ok.
